requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
docopt==0.6.2
idna==3.4
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==1.26.15
yarg==0.1.9
esy-osm-pbf==0.0.1